Tests
Warning: You have to change only the first parameter. I put the images in a for cycle, just for you. ;) The number in the filenames are the number of layers. You should run every file with every parameter number (1000, 2000, ...), 5 times. That's 3*6*5 = 90 runs. And save everything which was printed. In the meantime I will work on the better architecture. Since I can't run 2 networks at the same time, this architecture is up to you, but don't worry I will run more measurements with the next one. gl&hf!