Updated the GUI.
Separated the main container from its content, added a dummy icon and
created a seperated file for global settings.

Still need to add the checkbox for the image approximation as a
background and handle the proper resizing of the widgets when resizing
the window (so far, they just center in the extended window and that's
all). Potentially, add scrollbars to the layers' listboxes.